move data records into hook and out of context
