+++ src/app/map/[id]/providers/TableProvider.tsx
@@ -1,43 +1,12 @@
 "use client";
 
-import { keepPreviousData, useQuery } from "@tanstack/react-query";
 import { useCallback, useMemo, useState } from "react";
 import { TableContext } from "@/app/map/[id]/context/TableContext";
-import { useTRPC } from "@/services/trpc/react";
-import { useMapViews } from "../hooks/useMapViews";
 import type { ReactNode } from "react";
 
 const TableProvider = ({ children }: { children: ReactNode }) => {
-  const { view } = useMapViews();
-
   const [selectedDataSourceId, setSelectedDataSourceId] = useState<string>("");
   const [tablePage, setTablePage] = useState(0);
-
-  const dataSourceView = useMemo(
-    () =>
-      view?.dataSourceViews.find(
-        (dsv) => dsv.dataSourceId === selectedDataSourceId,
-      ),
-    [selectedDataSourceId, view?.dataSourceViews],
-  );
-
-  const trpc = useTRPC();
-
-  const dataRecordsQuery = useQuery(
-    trpc.dataRecord.list.queryOptions(
-      {
-        dataSourceId: selectedDataSourceId,
-        page: tablePage,
-        search: dataSourceView?.search,
-        filter: dataSourceView?.filter,
-        sort: dataSourceView?.sort,
-      },
-      {
-        enabled: Boolean(selectedDataSourceId),
-        placeholderData: keepPreviousData,
-      },
-    ),
-  );
 
   const handleDataSourceSelect = useCallback(
     (dataSourceId: string) => {
@@ -57,8 +26,6 @@ const TableProvider = ({ children }: { children: ReactNode }) => {
       selectedDataSourceId,
       setSelectedDataSourceId,
       handleDataSourceSelect,
-      dataRecordsResult: dataRecordsQuery.data,
-      dataRecordsLoading: dataRecordsQuery.isPending,
     };
   }, [
     tablePage,
@@ -66,8 +33,6 @@ const TableProvider = ({ children }: { children: ReactNode }) => {
     selectedDataSourceId,
     setSelectedDataSourceId,
     handleDataSourceSelect,
-    dataRecordsQuery.data,
-    dataRecordsQuery.isPending,
   ]);
   return <TableContext value={value}>{children}</TableContext>;
 };
